MyJuliaPackage

Documentation for MyJuliaPackage.

MyJuliaPackage.create_graph

MyJuliaPackage.explorer_1d

MyJuliaPackage.explorer_nd

MyJuliaPackage.gred

MyJuliaPackage.lowest_dur_neighbor

MyJuliaPackage.total_distance

MyJuliaPackage.create_graph — Method

Creates our own graph instance so that we don't have to recompute things so many times. :)

MyJuliaPackage.explorer_1d — Method

Base case for the recursion explorer. Checks just the best distance/duration in the set of streets we haven't traversed. :)

MyJuliaPackage.explorer_nd — Method

Explores arbitrary depth of the graph using recursion. Sums distance/duration to get the best overall speed. :)

MyJuliaPackage.gred — Method

Greedy search approach to returning Solution. Uses depth forward search to determine a greedy choice. :)

MyJuliaPackage.lowest_dur_neighbor — Method

Returns the neighboring (end, duration, distance) tuple with the lowest duration. :)

MyJuliaPackage.total_distance — Method

Returns the total distance of an array of itineraries